refactor option
